+++ app/src/main/java/seminar/outlawtrail/CalendarActivity.java
@@ -1,17 +1,16 @@
 package seminar.outlawtrail;
 
 import android.content.Intent;
-import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
 import android.view.View;
 
-public class AboutTrail extends AppCompatActivity {
+public class CalendarActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_about_trail);
-
+        setContentView(R.layout.activity_calendar);
     }
 
     public void go_home(View view) {
@@ -19,21 +18,20 @@ public class AboutTrail extends AppCompatActivity {
         startActivity(home);
     }
 
+    public void go_to_about(View view) {
+        Intent about = new Intent(this, AboutTrail.class);
+        startActivity(about);
+    }
+
     public void go_to_map(View view) {
         Intent map = new Intent(this, MapsActivity.class);
         startActivity(map);
-    }
-
-    public void go_to_calendar(View view) {
-        Intent calendar = new Intent(this, CalendarActivity.class);
-        startActivity(calendar);
     }
 
     public void go_to_contact(View view) {
         Intent contact = new Intent(this, ContactActivity.class);
         startActivity(contact);
     }
-
 
 
 }
