+++ app/src/main/java/seminar/outlawtrail/MainActivity.java
@@ -20,13 +20,22 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public void go_to_map(View view) {
-        Intent map = new Intent(this, WebMap.class);
+        Intent map = new Intent(this, MapsActivity.class);
         startActivity(map);
     }
 
     public void join_the_gang(View view){
         Intent joinGang = new Intent(Intent.ACTION_VIEW, Uri.parse("http://nebraskaoutlawtrail.org/join-the-gang-membership/"));
         startActivity(joinGang);
+    }
+    public void go_to_calendar(View view) {
+        Intent calendar = new Intent(this, CalendarActivity.class);
+        startActivity(calendar);
+    }
+
+    public void go_to_contact(View view) {
+        Intent contact = new Intent(this, ContactActivity.class);
+        startActivity(contact);
     }
