+++ app/src/main/java/seminar/outlawtrail/MapsActivity.java
@@ -3,22 +3,17 @@ package seminar.outlawtrail;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.webkit.WebView;
-import android.webkit.WebViewClient;
 
-public class WebMap extends AppCompatActivity {
-    WebView webView;
+public class MapsActivity extends AppCompatActivity {
     String mapPath = "https://www.google.com/maps/d/u/0/viewer?mid=z5cNDCbZgqkg.kReB3k9EGYyI";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_web_map);
+        setContentView(R.layout.activity_Maps);
 
     }
 
@@ -31,6 +26,17 @@ public class WebMap extends AppCompatActivity {
         Intent home = new Intent(this, MainActivity.class);
         startActivity(home);
     }
+
+    public void go_to_contact(View view) {
+        Intent contact = new Intent(this, ContactActivity.class);
+        startActivity(contact);
+    }
+
+    public void go_to_calendar(View view) {
+        Intent calendar = new Intent(this, CalendarActivity.class);
+        startActivity(calendar);
+    }
+
 
     public void launch_google(View view){
         Intent i = new Intent(Intent.ACTION_VIEW, Uri.parse(mapPath));
