+++ app/src/main/res/layout/activity_about_trail.xml
@@ -1,17 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto" android:layout_width="match_parent"
     android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    android:paddingBottom="@dimen/activity_vertical_margin" tools:context=".MainActivity"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    app:layout_behavior="@string/appbar_scrolling_view_behavior"
+    tools:showIn="@layout/activity_about" tools:context="seminar.outlawtrail.About"
     android:background="#bf956e">
+
 
     <LinearLayout
         android:orientation="horizontal"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:id="@+id/linearLayout"
+        android:id="@+id/linearLayout3"
         android:layout_alignParentTop="true"
         android:layout_centerHorizontal="true">
 
@@ -19,37 +23,36 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="Home"
-            android:id="@+id/homeButton" />
+            android:id="@+id/button" />
 
         <Button
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="About the Trail"
-            android:id="@+id/aboutButton"
-            android:onClick="go_to_about" />
+            android:id="@+id/button2"
+             />
 
         <Button
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="Map"
-            android:id="@+id/button4"
+            android:id="@+id/button3"
             android:onClick="go_to_map" />
-
     </LinearLayout>
 
     <LinearLayout
         android:orientation="horizontal"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:id="@+id/linearLayout2"
-        android:layout_below="@+id/linearLayout"
+        android:id="@+id/linearLayout4"
+        android:layout_below="@+id/linearLayout3"
         android:layout_centerHorizontal="true">
 
         <Button
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="Calendar"
-            android:id="@+id/button5"
+            android:id="@+id/button6"
             android:layout_below="@+id/linearLayout"
             android:layout_alignEnd="@+id/joinButton"
             android:onClick="go_to_calendar" />
@@ -58,7 +61,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="Contact"
-            android:id="@+id/button12"
+            android:id="@+id/button14"
             android:layout_below="@+id/linearLayout"
             android:layout_alignEnd="@+id/joinButton"
             android:onClick="go_to_contact" />
@@ -68,50 +71,63 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:textAppearance="?android:attr/textAppearanceLarge"
-        android:text="Escape to Nebraska’s Scenic Outlaw Trail for your next vacation!"
-        android:id="@+id/textView"
+        android:text="Are you ready to explore the Outlaw Trail Scenic Byway 12?"
+        android:id="@+id/textView8"
         android:textAlignment="center"
-        android:layout_below="@+id/linearLayout2"
+        android:layout_below="@+id/linearLayout4"
         android:layout_centerHorizontal="true" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textAppearance="?android:attr/textAppearanceMedium"
-        android:text="Journey from the seemingly endless fields of fertile farm land in Eastern Nebraska, to the grass covered sand dunes known as the Sandhills."
-        android:id="@+id/textView10"
-        android:textAlignment="center"
-        android:textColor="#000000"
-        android:layout_below="@+id/textView20"
-        android:layout_alignParentStart="true" />
-
-    <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Join The Gang"
-        android:id="@+id/joinButton"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:onClick="join_the_gang" />
 
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:textAppearance="?android:attr/textAppearanceSmall"
-        android:text="Will send to external link on website to signup for support"
-        android:id="@+id/textView3"
+        android:text="Once  traveled by famous outlaws of the day, including Jesse James and his gang , Kid Wade, and Doc Middleton, Nebraska’s Outlaw Trail Scenic Byway 12 features rich history, exciting events,  and breathtaking scenery."
+        android:id="@+id/textView9"
         android:textAlignment="center"
-        android:layout_above="@+id/joinButton"
+        android:textColor="#000000"
+        android:layout_below="@+id/textView8"
+        android:layout_alignParentStart="true" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceLarge"
+        android:text="Scenic Byway"
+        android:id="@+id/textView4"
+        android:textAlignment="center"
+        android:layout_below="@+id/textView9"
         android:layout_centerHorizontal="true" />
 
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:textAppearance="?android:attr/textAppearanceMedium"
-        android:id="@+id/textView20"
+        android:textAppearance="?android:attr/textAppearanceSmall"
+        android:text="From Valentine, to South Sioux City, this scenic trail encompasses 231 miles of pristine beauty and undiscovered treasures."
+        android:id="@+id/textView21"
         android:textAlignment="center"
         android:textColor="#000000"
-        android:layout_below="@+id/textView"
+        android:layout_below="@+id/textView4"
         android:layout_centerHorizontal="true" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceLarge"
+        android:text="Recreation"
+        android:id="@+id/textView7"
+        android:textAlignment="center"
+        android:layout_below="@+id/textView21"
+        android:layout_centerHorizontal="true" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceSmall"
+        android:text="Popular outdoor family activities include trail rides, camping, golfing, swimming, boating, hiking, fishing, hunting, wagon rides, and river activities, including tubing and canoeing on the Niobrara River."
+        android:id="@+id/textView5"
+        android:textAlignment="center"
+        android:textColor="#000000"
+        android:layout_below="@+id/textView7"
+        android:layout_alignParentStart="true" />
 
 </RelativeLayout>
