+++ app/src/main/res/layout/activity_calendar.xml
@@ -51,67 +51,34 @@
             android:text="Calendar"
             android:id="@+id/button5"
             android:layout_below="@+id/linearLayout"
-            android:layout_alignEnd="@+id/joinButton"
-            android:onClick="go_to_calendar" />
+            android:layout_alignEnd="@+id/joinButton" />
 
         <Button
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="Contact"
-            android:id="@+id/button12"
+            android:id="@+id/button13"
             android:layout_below="@+id/linearLayout"
             android:layout_alignEnd="@+id/joinButton"
             android:onClick="go_to_contact" />
     </LinearLayout>
 
-    <TextView
+    <CalendarView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:textAppearance="?android:attr/textAppearanceLarge"
-        android:text="Escape to Nebraska’s Scenic Outlaw Trail for your next vacation!"
-        android:id="@+id/textView"
-        android:textAlignment="center"
+        android:id="@+id/calendarView"
+        android:background="#bf956e"
         android:layout_below="@+id/linearLayout2"
         android:layout_centerHorizontal="true" />
 
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:textAppearance="?android:attr/textAppearanceMedium"
-        android:text="Journey from the seemingly endless fields of fertile farm land in Eastern Nebraska, to the grass covered sand dunes known as the Sandhills."
-        android:id="@+id/textView10"
-        android:textAlignment="center"
-        android:textColor="#000000"
-        android:layout_below="@+id/textView20"
-        android:layout_alignParentStart="true" />
-
-    <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Join The Gang"
-        android:id="@+id/joinButton"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:onClick="join_the_gang" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
         android:textAppearance="?android:attr/textAppearanceSmall"
-        android:text="Will send to external link on website to signup for support"
-        android:id="@+id/textView3"
+        android:text="This is not functional, but at least have an option of keeping people in the loop with a calendar application like this."
+        android:id="@+id/textView2"
         android:textAlignment="center"
-        android:layout_above="@+id/joinButton"
-        android:layout_centerHorizontal="true" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textAppearance="?android:attr/textAppearanceMedium"
-        android:id="@+id/textView20"
-        android:textAlignment="center"
-        android:textColor="#000000"
-        android:layout_below="@+id/textView"
+        android:layout_below="@+id/calendarView"
         android:layout_centerHorizontal="true" />
 
 </RelativeLayout>
