+++ app/src/main/res/layout/activity_contact.xml
@@ -1,0 +1,199 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
+    android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingRight="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
+    android:paddingBottom="@dimen/activity_vertical_margin" tools:context=".MainActivity"
+    android:background="#bf956e">
+
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/linearLayout"
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true">
+
+        <Button
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Home"
+            android:id="@+id/homeButton" />
+
+        <Button
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="About the Trail"
+            android:id="@+id/aboutButton"
+            android:onClick="go_to_about" />
+
+        <Button
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Map"
+            android:id="@+id/button4"
+            android:onClick="go_to_map" />
+
+    </LinearLayout>
+
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/linearLayout2"
+        android:layout_below="@+id/linearLayout"
+        android:layout_centerHorizontal="true">
+
+        <Button
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Calendar"
+            android:id="@+id/button5"
+            android:layout_below="@+id/linearLayout"
+            android:layout_alignEnd="@+id/joinButton"
+            android:onClick="go_to_calendar" />
+
+        <Button
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Contact"
+            android:id="@+id/button12"
+            android:layout_below="@+id/linearLayout"
+            android:layout_alignEnd="@+id/joinButton" />
+    </LinearLayout>
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceLarge"
+        android:text="Feel free to connect with us!"
+        android:id="@+id/textView"
+        android:textAlignment="center"
+        android:layout_below="@+id/linearLayout2"
+        android:layout_centerHorizontal="true" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceMedium"
+        android:id="@+id/textView17"
+        android:textAlignment="center"
+        android:textColor="#000000"
+        android:layout_below="@+id/textView"
+        android:layout_alignStart="@+id/linearLayout2"
+        android:layout_alignEnd="@+id/textView10" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceMedium"
+        android:text="Rhonda Kneifl, Pres."
+        android:id="@+id/textView10"
+        android:textAlignment="center"
+        android:textColor="#000000"
+        android:layout_below="@+id/textView17"
+        android:layout_alignStart="@+id/textView12"
+        android:layout_alignLeft="@+id/linearLayout" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceMedium"
+        android:text="     PO Box 321 Newcastle, NE 68757 "
+        android:id="@+id/textView12"
+        android:textAlignment="center"
+        android:textColor="#000000"
+        android:layout_below="@+id/textView10"
+        android:layout_alignStart="@+id/linearLayout" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceMedium"
+        android:text="     402-322-3360"
+        android:id="@+id/textView11"
+        android:textAlignment="center"
+        android:textColor="#000000"
+        android:layout_below="@+id/textView12"
+        android:layout_alignLeft="@+id/linearLayout" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceMedium"
+        android:text="Marci Broyhill, Sec. "
+        android:id="@+id/textView13"
+        android:textAlignment="center"
+        android:textColor="#000000"
+        android:layout_below="@+id/textView11"
+        android:layout_alignStart="@+id/textView11"
+        android:layout_alignLeft="@+id/linearLayout" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceMedium"
+        android:text="     PO Box 648 Dakota City, NE 68731"
+        android:id="@+id/textView14"
+        android:textAlignment="center"
+        android:textColor="#000000"
+        android:layout_below="@+id/textView13"
+        android:layout_alignStart="@+id/textView13" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceMedium"
+        android:text="     402-987-3413 "
+        android:id="@+id/textView15"
+        android:textAlignment="center"
+        android:textColor="#000000"
+        android:layout_below="@+id/textView14"
+        android:layout_alignStart="@+id/textView14" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceMedium"
+        android:text="Facebook.com/"
+        android:id="@+id/textView16"
+        android:textAlignment="center"
+        android:textColor="#000000"
+        android:layout_below="@+id/textView15"
+        android:layout_alignStart="@+id/textView15"
+        android:layout_alignLeft="@+id/linearLayout" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceMedium"
+        android:text="     outlawtrailscenicbyway"
+        android:id="@+id/textView18"
+        android:textAlignment="center"
+        android:textColor="#000000"
+        android:layout_below="@+id/textView16"
+        android:layout_alignStart="@+id/textView15" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceMedium"
+        android:text="plus.google.com/u/0/106336481720792541207/posts"
+        android:id="@+id/textView19"
+        android:textColor="#000000"
+        android:layout_below="@+id/textView6"
+        android:layout_alignStart="@+id/textView" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceMedium"
+        android:text="Google+"
+        android:id="@+id/textView6"
+        android:textAlignment="center"
+        android:textColor="#000000"
+        android:layout_below="@+id/textView18"
+        android:layout_alignStart="@+id/textView18" />
+
+</RelativeLayout>
