+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,6 @@
     <string name="title_activity_maps">Map</string>
     <string name="title_activity_about">About</string>
     <string name="title_activity_web_map">WebMap</string>
+    <string name="title_activity_contact">ContactActivity</string>
+    <string name="title_activity_calendar">CalendarActivity</string>
 </resources>
